@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "Unified backend for HabitForge",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "start": "node index.js",
     "dev": "nodemon index.js"
@@ -11,10 +12,10 @@
     "express": "^4.18.2",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
-    "@google/genai": "^0.1.1",
+    "@google/genai": "*",
     "nodemailer": "^6.9.1"
   },
   "engines": {
-    "node": ">=16.0.0"
+    "node": ">=18.0.0"
   }
 }
